test(brave): split client tests into http_tests + classify_tests
brave/client/tests.rs (465) split into sibling http_tests.rs (367) and
classify_tests.rs (91), matching the existing fetch/ssrf/{tests,dns_tests}.rs
layout. Both files are now <=400 lines. Production client.rs is unchanged
apart from the mod declaration; all 449 tests pass before and after.

ADR-0007 Confirmation reference updated to brave/client/http_tests.rs to
track the moved module.
